stupid hack: source bashrc.local in the bash_completion dir...
This way makes sure all the custom crap gets invoked whether or not we are in a login shell. It is also better than /etc/profile.d because \(at least for now\) the bash_completion stuff gets sourced at the END of /etc/bash.bashrc, after all the stupid ugly PATH and TERM variables are reset.