[encrypt_ppz] load detached cryptsetup headers from another device
There is a bizarre bug where cryptsetup cannot load detached headers with "--header $file" when $file is on the initram filesystem. For some reason it works when $file is on a mounted device. I have no idea why.

As a workaround, this lets me pass "headerdev=/dev/foo" to the kernel to make cryptsetup look on an external device automatically.