ditch python-ntp
it's a nice idea but never in all these years have i both seen it fail and acted on that.
ntpd works fine and systemd-journal makes it easy to keep tabs on it.
having this in integrity also leaked information to the network on what i'm doing, which is minor but never a good thing.